+++ packages/ses/test/test-scope-constants.js
@@ -1,59 +1,95 @@
 import test from 'ava';
 import { getScopeConstants } from '../src/scope-constants.js';
 
-test('getScopeConstants - global object', t => {
+test('getScopeConstants - globalObject', t => {
   t.plan(20);
 
   t.deepEqual(
     getScopeConstants({}),
-    { globalLexicalConstants: [], globalObjectConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalLexicalConstants: [],
+      globalObjectConstants: [],
+    },
     'should return empty if no global',
   );
 
   t.deepEqual(
     getScopeConstants({ foo: true }),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject configurable & writable',
   );
   t.deepEqual(
     getScopeConstants(Object.create(null, { foo: { value: true } })),
-    { globalObjectConstants: ['foo'], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: ['foo'],
+      globalLexicalConstants: [],
+    },
     'should return non configurable & non writable',
   );
   t.deepEqual(
     getScopeConstants(
       Object.create(null, { foo: { value: true, configurable: true } }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject configurable',
   );
   t.deepEqual(
     getScopeConstants(
       Object.create(null, { foo: { value: true, writable: true } }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject writable',
   );
 
   t.deepEqual(
     getScopeConstants(Object.create(null, { foo: { get: () => true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject getter',
   );
   t.deepEqual(
     getScopeConstants(Object.create(null, { foo: { set: () => true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject setter',
   );
 
   t.deepEqual(
     getScopeConstants(Object.create(null, { eval: { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject eval',
   );
   t.deepEqual(
     getScopeConstants(Object.create(null, { const: { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject reserved keyword',
   );
   t.deepEqual(
@@ -64,7 +100,11 @@ test('getScopeConstants - global object', t => {
         false: { value: true },
       }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject literals (reserved)',
   );
   t.deepEqual(
@@ -74,46 +114,78 @@ test('getScopeConstants - global object', t => {
         arguments: { value: true },
       }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject this and arguments',
   );
   t.deepEqual(
     getScopeConstants(
       Object.create(null, { [Symbol.iterator]: { value: true } }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject symbols',
   );
 
   t.deepEqual(
     getScopeConstants(Object.create(null, { 123: { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject leading digit',
   );
   t.deepEqual(
     getScopeConstants(Object.create(null, { '-123': { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject leading dash',
   );
 
   t.deepEqual(
     getScopeConstants(Object.create(null, { _123: { value: true } })),
-    { globalObjectConstants: ['_123'], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: ['_123'],
+      globalLexicalConstants: [],
+    },
     'should return leading underscore',
   );
   t.deepEqual(
     getScopeConstants(Object.create(null, { $123: { value: true } })),
-    { globalObjectConstants: ['$123'], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: ['$123'],
+      globalLexicalConstants: [],
+    },
     'should return leading underscore',
   );
   t.deepEqual(
     getScopeConstants(Object.create(null, { a123: { value: true } })),
-    { globalObjectConstants: ['a123'], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: ['a123'],
+      globalLexicalConstants: [],
+    },
     'should return leading lowercase',
   );
   t.deepEqual(
     getScopeConstants(Object.create(null, { A123: { value: true } })),
-    { globalObjectConstants: ['A123'], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: ['A123'],
+      globalLexicalConstants: [],
+    },
     'should return leading uppercase',
   );
 
@@ -121,7 +193,11 @@ test('getScopeConstants - global object', t => {
     getScopeConstants(
       Object.create(null, { foo: { value: true }, bar: { value: true } }),
     ),
-    { globalObjectConstants: ['foo', 'bar'], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: ['foo', 'bar'],
+      globalLexicalConstants: [],
+    },
     'should return all non configurable & non writable',
   );
   t.deepEqual(
@@ -131,28 +207,44 @@ test('getScopeConstants - global object', t => {
         bar: { value: true, configurable: true },
       }),
     ),
-    { globalObjectConstants: ['foo'], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: ['foo'],
+      globalLexicalConstants: [],
+    },
     'should return only non configurable & non writable',
   );
 });
 
-test('getScopeConstants - local object (endownments)', t => {
+test('getScopeConstants - globalLexicals', t => {
   t.plan(20);
 
   t.deepEqual(
     getScopeConstants({}, {}),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
-    'should return empty if no local',
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
+    'should return empty if no globalLexicals',
   );
 
   t.deepEqual(
     getScopeConstants({}, { foo: true }),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject configurable & writable',
   );
   t.deepEqual(
     getScopeConstants({}, Object.create(null, { foo: { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: ['foo'] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: ['foo'],
+    },
     'should return non configurable & non writable',
   );
   t.deepEqual(
@@ -160,7 +252,11 @@ test('getScopeConstants - local object (endownments)', t => {
       {},
       Object.create(null, { foo: { value: true, configurable: true } }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject configurable',
   );
   t.deepEqual(
@@ -168,29 +264,49 @@ test('getScopeConstants - local object (endownments)', t => {
       {},
       Object.create(null, { foo: { value: true, writable: true } }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject writable',
   );
 
   t.deepEqual(
     getScopeConstants({}, Object.create(null, { foo: { get: () => true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject getter',
   );
   t.deepEqual(
     getScopeConstants({}, Object.create(null, { foo: { set: () => true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject setter',
   );
 
   t.deepEqual(
     getScopeConstants({}, Object.create(null, { eval: { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject eval',
   );
   t.deepEqual(
     getScopeConstants({}, Object.create(null, { const: { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject reserved keyword',
   );
   t.deepEqual(
@@ -202,7 +318,11 @@ test('getScopeConstants - local object (endownments)', t => {
         false: { value: true },
       }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject literals (reserved)',
   );
   t.deepEqual(
@@ -213,7 +333,11 @@ test('getScopeConstants - local object (endownments)', t => {
         arguments: { value: true },
       }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject this and arguments',
   );
   t.deepEqual(
@@ -221,39 +345,67 @@ test('getScopeConstants - local object (endownments)', t => {
       {},
       Object.create(null, { [Symbol.iterator]: { value: true } }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject symbols',
   );
 
   t.deepEqual(
     getScopeConstants({}, Object.create(null, { 123: { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject leading digit',
   );
   t.deepEqual(
     getScopeConstants({}, Object.create(null, { '-123': { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: [] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: [],
+    },
     'should reject leading dash',
   );
 
   t.deepEqual(
     getScopeConstants({}, Object.create(null, { _123: { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: ['_123'] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: ['_123'],
+    },
     'should return leading underscore',
   );
   t.deepEqual(
     getScopeConstants({}, Object.create(null, { $123: { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: ['$123'] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: ['$123'],
+    },
     'should return leading underscore',
   );
   t.deepEqual(
     getScopeConstants({}, Object.create(null, { a123: { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: ['a123'] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: ['a123'],
+    },
     'should return leading lowercase',
   );
   t.deepEqual(
     getScopeConstants({}, Object.create(null, { A123: { value: true } })),
-    { globalObjectConstants: [], globalLexicalConstants: ['A123'] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: ['A123'],
+    },
     'should return leading uppercase',
   );
 
@@ -262,7 +414,11 @@ test('getScopeConstants - local object (endownments)', t => {
       {},
       Object.create(null, { foo: { value: true }, bar: { value: true } }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: ['foo', 'bar'] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: ['foo', 'bar'],
+    },
     'should return all non configurable & non writable',
   );
   t.deepEqual(
@@ -273,12 +429,16 @@ test('getScopeConstants - local object (endownments)', t => {
         bar: { value: true, configurable: true },
       }),
     ),
-    { globalObjectConstants: [], globalLexicalConstants: ['foo'] },
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: [],
+      globalLexicalConstants: ['foo'],
+    },
     'should return only non configurable & non writable',
   );
 });
 
-test('getScopeConstants - global and local object', t => {
+test('getScopeConstants - globalObject and globalLexicals', t => {
   t.plan(1);
 
   t.deepEqual(
@@ -286,7 +446,11 @@ test('getScopeConstants - global and local object', t => {
       Object.create(null, { foo: { value: true }, bar: { value: true } }),
       { foo: false },
     ),
-    { globalObjectConstants: ['bar'], globalLexicalConstants: [] },
-    'should only return global contants not hidden by local',
+    {
+      moduleLexicalConstants: [],
+      globalObjectConstants: ['bar'],
+      globalLexicalConstants: [],
+    },
+    'should only return global contants not hidden by global lexicals',
   );
 });